requirements.txt
fastapi
uvicorn
sqlmodel
aiosqlite
pydantic
pydantic-settings
python-jose
passlib
bcrypt
msal
httpx
python-multipart
email-validator
python-dotenv
PyJWT
cryptography
alembic
jinja2
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
starlette-exporter
redis
pytest